backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.2,<3
python-dotenv==1.0.0
openai==1.3.7
pydantic==2.5.0
python-multipart==0.0.6
yfinance==0.2.32
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
pydantic-settings==2.1.0
